[Change] Full remake by claude 4.6 (client,server,shared lib,tools,docs)

+++ tools/bot/BotInterface.cpp
@@ -6,7 +6,7 @@ void BotInterface::dropAllPlayerOnTheMap(CatchChallenger::Api_protocol_Qt *api)
 }
 
 void BotInterface::insert_player_all(CatchChallenger::Api_protocol_Qt *api, const CatchChallenger::Player_public_informations &player,
-                const quint32 &mapId, const quint16 &x, const quint16 &y, const CatchChallenger::Direction &direction)
+                const uint8_t &mapId, const uint8_t &x, const uint8_t &y, const CatchChallenger::Direction &direction)
 {
     (void)api;
     (void)player;
@@ -16,7 +16,7 @@ void BotInterface::insert_player_all(CatchChallenger::Api_protocol_Qt *api, cons
     (void)direction;
 }
 
-void BotInterface::remove_player(CatchChallenger::Api_protocol_Qt *api, const uint16_t &id)
+void BotInterface::remove_player(CatchChallenger::Api_protocol_Qt *api, const uint8_t &id)
 {
     (void)api;
     (void)id;
