+++ tools/bot/simple/SimpleAction.h
@@ -18,7 +18,7 @@ public:
     SimpleAction();
     ~SimpleAction();
     void insert_player(CatchChallenger::Api_protocol_Qt *api,const CatchChallenger::Player_public_informations &player,
-                       const uint32_t &mapId,const uint16_t &x,const uint16_t &y,const CatchChallenger::Direction &direction);
+                       const uint8_t &mapId,const uint8_t &x,const uint8_t &y,const CatchChallenger::Direction &direction);
 private:
     QTimer moveTimer;
     QTimer textTimer;
