+++ tools/bot/simple/SimpleBotInterface.cpp
@@ -58,7 +58,7 @@ QStringList SimpleBotInterface::variablesList()
     return QStringList() << QString("move") << QString("randomText") << QString("bugInDirection") << QString("globalChatRandomReply");
 }
 
-void SimpleBotInterface::insert_player(CatchChallenger::Api_protocol_Qt *api,const CatchChallenger::Player_public_informations &player,const quint32 &mapId,const quint16 &x,const quint16 &y,const CatchChallenger::Direction &direction)
+void SimpleBotInterface::insert_player(CatchChallenger::Api_protocol_Qt *api,const CatchChallenger::Player_public_informations &player,const uint8_t &mapId,const uint8_t &x,const uint8_t &y,const CatchChallenger::Direction &direction)
 {
     Player newPlayer;
     newPlayer.player=player;
